Speed up font watching loop for checking test spans
The original value was somewhat arbitrary and some browsers can do the check
faster than this if we set the timeout faster. This is still probably slow
enough not to put undue burden on browsers and it negates any speed loss from
the previous change of needing to see a new size twice before we trigger
active.