Remove UserAgent from DomHelper, replace with feature detection
UserAgent was required to construct a DomHelper, but the only thing it was
used for is in setStyle to determine if we're in an old IE and need to use a
different approach. I was using DomHelper in some code and finding it cumbersome
to have a UserAgent instance around as well. This goal is better accomplished
with a feature-detection approach anyway.

I've replaced the UserAgent arg with feature detection for getting/setting the
style attribute. This is a very similar approach to the one that jQuery uses to
detect the same feature.

This has the added side-effect of allowing many of the modules to no longer
construct a UserAgent instance using the parser, since they only used it for the
DomHelper, which simplifies things.