Initial code to combine bug and last resort size detection (needs test fixes)
Because we need to know the last resort sizes to ignore, and the last resort
sizes can depend on the font stack and text content, we need to actually detect
them in the same context to be 100% sure that they'll be the same. This change
creates more null fonts on the page (one for each variation that we're
detecting), but it seems like these don't stay behind in the console, and they
hopefully shouldn't provide too much of a drain performance wise. If they do,
we can do additional work to use a single global null font rule for every
round of watchers (but since they're time-limited before the browser detects
that they don't contain a valid font, we'll need to make a new one for each
round of detection).