Change README instances of "WebFont Loader" to "Web Font Loader"
Both Google and Typekit have standardized on using "web fonts" instead of a
single "webfonts" to refer to fonts included with font-face on the page. This
awkward capitalization was originally written back when the industry was still
figuring this out, but we should make it consistent with our preferred
terminology at this point. Luckily, nothing in code needs to be changed since
CamelCase works the same way no matter what.